Add and persist city programmatically + Start to implement Maps

+++ app/src/main/java/com/borombo/mobileassignment/model/City.java
@@ -7,8 +7,8 @@ package com.borombo.mobileassignment.model;
 public class City {
 
     private String name;
-    private Long latitude;
-    private Long longitude;
+    private Double latitude;
+    private Double longitude;
 
 
     public String getName() {
@@ -19,19 +19,19 @@ public class City {
         this.name = name;
     }
 
-    public Long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public Long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
 }
